Completely reinvented the pipelining system, which now allows for OpenGL-like use, as well as the standard Vulkan usage (which gives out a little better performance), added support for both compiled and non-compiled shaders, made automatic shader compilation through CMake and Python, made the Engine part almost completely standalone (does not directly interact with the Core rendering logic, made all dependencies statically linked and temporarily removed the shadows feature

+++ Core/Rendering/UI/ImGuiCore.cpp
@@ -96,7 +96,7 @@ namespace Sierra::Core::Rendering::UI
 
     /* --- POLLING METHODS --- */
 
-    bool ImGuiCore::SetSceneViewSize(const uint newWidth, const uint newHeight)
+    bool ImGuiCore::SetSceneViewSize(const float newWidth, const float newHeight)
     {
         if (instance.sceneViewWidth == newWidth && instance.sceneViewHeight == newHeight) return false;
